fix(x86_64): boot NanOS on real UEFI hardware (Dell Latitude 5310) from USB
Five real-hardware-only bugs that QEMU/OVMF had hidden, found via a
framebuffer POST-code ladder (kept arch::debugBar as a bring-up primitive):

1. ESP must be FAT32, not FAT16. mtools picked FAT16 at 32 MiB and the Dell
   firmware silently reset instead of reading the ESP. 64 MiB ESP + mformat -F;
   root moved 34->66 MiB (ROOT_OFFSET + Makefile IMAGE64_GRUB2_PART offset).

2. loader.S identity-mapped only 1 GiB; on 8-16 GiB machines the Multiboot info
   / E820 mmap / framebuffer sit higher and a pre-initPaging access faulted.
   Map [0,64 GiB) with 2 MiB pages (no pdpe1gb, QEMU-safe).

3. Framebuffer/MMIO physical address truncated to 32-bit (mmuMapKernelMmio,
   mmuMapUserFb, FbInfo.phys); real-HW LFB/BARs sit >4 GiB. Widened to 64-bit.

4. mmuInitKernel huge-maps [0,topOfRam) then mapRange'd the LAPIC + framebuffer
   4 KiB pages. On real HW those MMIO addresses fall INSIDE the huge map, so
   mapRange walked the live 2 MiB huge PDE as a page-table pointer -> corruption
   -> triple fault. Guard with g_identityTop: skip MMIO already huge-mapped.

5. USB-MSC multi-sector reads returned only the first 512 B on real xHCI, so ext
   directory blocks read back half-zero and init was "not found". One READ(10)
   per sector (mirrors the ATA multi-sector fix). Also set Port Power on every
   xHCI port + poll for connect debounce before enumerating (real ports start
   unpowered; QEMU powered them and set CCS instantly).

+++ arch/include/arch/mmu.h
@@ -48,7 +48,11 @@ uint32_t mmuKernelDirPhys();
 // RAM and is therefore NOT covered by the kernel identity map) into the kernel
 // directory, present+writable, supervisor. Call after mmuInitKernel and BEFORE any
 // per-process address space is created, so the new PDE is shared by every process.
-void mmuMapKernelMmio(uint32_t phys, uint32_t bytes);
+// phys is 64-bit: PCI MMIO BARs (framebuffer, NIC, xHCI) on real x86_64 hardware are
+// routinely programmed above 4 GiB. Truncating to 32 bits maps the wrong page and the
+// first access triple-faults on the metal (QEMU/OVMF places them low, hiding it). The
+// i686 impl narrows internally (its physical space is 32-bit anyway).
+void mmuMapKernelMmio(uint64_t phys, uint32_t bytes);
 
 // Read/load the active page-directory physical address (CR3 on x86). MI code uses
 // these to stage a spawned child under the kernel identity map, then restore the
@@ -76,7 +80,7 @@ uint32_t mmuSpaceDirPhys(AddressSpace*);
 // Map a framebuffer's physical region into a process address space at a fixed user VA
 // (above RAM, separate from the 1 MiB user window), present+writable+user. Returns the
 // user virtual address of the framebuffer, or 0 on failure. Used by mmap of /dev/fb0.
-uint32_t mmuMapUserFb(AddressSpace*, uint32_t fbPhys, uint32_t bytes);
+uint32_t mmuMapUserFb(AddressSpace*, uint64_t fbPhys, uint32_t bytes);   // fbPhys 64-bit: real HW LFB >4 GiB
 
 // Growable anonymous user heap (the brk/sbrk region). It lives at a fixed high VA,
 // above RAM and the framebuffer window, so it is independent of the 4 MiB user window.
